Complete Assignment 9 and update  "Design" file

+++ Lab02/AimsProject/DigitalVideoDisc.java
@@ -47,7 +47,7 @@ public class DigitalVideoDisc {
         return length;
     }
 
-    public double getCost() {
+    public float getCost() {
         return cost;
     }
